Split into Multiple Scripts; Small Cleanup
