+++ overrides/groovy/classes/postInit/common.groovy
@@ -9,18 +9,19 @@ class Common {
 	private static ItemStack meP2p = null
 	private static List<Pair<String, ItemStack>> p2pVariants = null
 	private static List<ItemStack> eioGlasses = null
-	private static List<String> voltageNames = null
+	private static List<Pair<Integer, String>> voltageNames = null
 
 	/**
-	 * Gets a continuous sublist of GregTech Voltage Names, in lowercase. Do not modify the returned list.
+	 * Gets a continuous sublist of GregTech tier to voltage names, in lowercase.<br>
+	 * <b>Do not modify the returned list!</b>
 	 * @param from The start index of the sublist. Inclusive.
 	 * @param to The end index of the sublist. Inclusive.
  	 */
-	static List<String> getVoltageNames(int from, int to) {
+	static List<Pair<Integer, String>> getVoltageNames(int from, int to) {
 		if (voltageNames == null) {
 			voltageNames = new ArrayList<>()
-			for (var voltage : GTValues.VN) {
-				voltageNames.add(voltage.toLowerCase())
+			for (int i = 0; i < GTValues.VN.length; i++) {
+				voltageNames.add(Pair.of(i, GTValues.VN[i].toLowerCase()))
 			}
 		}
